# Much faster CPU prompt processing (part 2)