@Nexesenex 2025-06-17
Very impressive, @ikawrakow!
All your recent commits motivates me to put more of IK_Llama on my Kobold.Cpp fork.
I already have overall twice its CPU PP perfs thanks to your amazing work, and I merged most of your quants, including the last Trellis!
Way to make an enthusiast happy!